Moved Render Settings to its own class
