+++ SPLASH/src/game/states/PBRTestState.cpp
@@ -67,6 +67,8 @@ PBRTestState::PBRTestState(StateStack& stack)
 	// Create octree
 	m_octree = SAIL_NEW Octree(nullptr);
 
+	m_renderSettingsWindow.activateMaterialPicking(&m_cam, m_octree);
+
 	// Get the Application instance
 	m_app = Application::getInstance();
 	m_componentSystems.renderSystem = ECS::Instance()->getSystem<RenderSystem>();
@@ -263,54 +265,9 @@ bool PBRTestState::render(float dt, float alpha) {
 bool PBRTestState::renderImgui(float dt) {
 	// The ImGui window is rendered when activated on F10
 	ImGui::ShowDemoWindow();
-	renderImGuiRenderSettings(dt);
 	renderImGuiLightDebug(dt);
 	m_profiler.renderWindow();
-
-	return false;
-}
-
-bool PBRTestState::renderImGuiRenderSettings(float dt) {
-	ImGui::Begin("Rendering settings");
-	ImGui::Checkbox("Enable post processing",
-		&(*Application::getInstance()->getRenderWrapper()).getDoPostProcessing()
-	);
-
-	static Entity* pickedEntity = nullptr;
-	static float metalness = 1.0f;
-	static float roughness = 1.0f;
-	static float ao = 1.0f;
-
-	ImGui::Separator();
-	if (ImGui::Button("Pick entity")) {
-		Octree::RayIntersectionInfo tempInfo;
-		m_octree->getRayIntersection(m_cam.getPosition(), m_cam.getDirection(), &tempInfo);
-		if (tempInfo.closestHitIndex != -1) {
-			pickedEntity = tempInfo.info.at(tempInfo.closestHitIndex).entity;
-		}
-	}
-
-	if (pickedEntity) {
-		ImGui::Text("Material properties for %s", pickedEntity->getName());
-		if (auto * model = pickedEntity->getComponent<ModelComponent>()) {
-			auto* mat = model->getModel()->getMesh(0)->getMaterial();
-			const auto& pbrSettings = mat->getPBRSettings();
-			metalness = pbrSettings.metalnessScale;
-			roughness = pbrSettings.roughnessScale;
-			ao = pbrSettings.aoScale;
-			if (ImGui::SliderFloat("Metalness scale", &metalness, 0.f, 1.f)) {
-				mat->setMetalnessScale(metalness);
-			}
-			if (ImGui::SliderFloat("Roughness scale", &roughness, 0.f, 1.f)) {
-				mat->setRoughnessScale(roughness);
-			}
-			if (ImGui::SliderFloat("AO scale", &ao, 0.f, 1.f)) {
-				mat->setAOScale(ao);
-			}
-		}
-	}
-
-	ImGui::End();
+	m_renderSettingsWindow.renderWindow();
 
 	return false;
 }
