+++ SPLASH/src/game/states/PerformanceTestState.h
@@ -28,7 +28,6 @@ public:
 
 private:
 	bool onResize(WindowResizeEvent& event);
-	bool renderImGuiRenderSettings(float dt);
 	bool renderImGuiLightDebug(float dt);
 
 private:
@@ -51,6 +50,7 @@ private:
 	Systems m_componentSystems;
 	LightSetup m_lights;
 	Profiler m_profiler;
+	RenderSettingsWindow m_renderSettingsWindow;
 
 	size_t m_currLightIndex;
 	// For use by non-deterministic entities
